Move patternfly toolbar configs to external type file.
